pco-ai-service/requirements.txt
# PCO AI Service - Python Dependencies

# Core dependencies
flask>=2.3.0              # Web framework for REST API
python-dotenv>=1.0.0      # Environment variable management
requests>=2.31.0          # HTTP library for API calls

# Development dependencies (optional)
pytest>=7.4.0             # Testing framework
pytest-cov>=4.1.0         # Test coverage
black>=23.7.0             # Code formatter
flake8>=6.1.0             # Linter